requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
idna==2.10
mysql-connector-python==8.0.25
mysqlclient==2.0.3
protobuf==3.17.2
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
